Specify a v5 key version and a new fingerprint scheme.
The v5 key version is introduced to
  a) to trigger the use of the new fingerprint scheme,
  b) to prepare for algorithms which need keys larger than 64k,
  c) to ease parsing of unknown algorithms.

The fingerprint algorithm uses SHA-256 because
  a) 32 octets are sufficient for a fingerprint
     (#include "640k-ram-will-always-be-enough.joke"),
  b) SHA-256 is well matured and widely available,
  c) SHA-256 is faster than SHA-512 on embedded platforms,
  d) implementations need to support SHA-256 anyway because it is
     the commonly used hash algorithms for signatures.

Although the fingerprint is specified at full length it is truncated
to 25 octets for purposes of the OpenPGP spec.  This is so that
signatures are not too much enlarged without a good reason.

A human readable representation of the fingerprint is not given
because that was never done in OpenPGP.  Implementations may for
example use

  1122334455 6677889900 aabbccddee ff00112233 4455667788
or
  11223 34455 66778 89900 aabbc cddee ff001 12233 44556 67788

to show fingerprints.

The Key ID is still defined because the 64 bits do not pose a problem
when used to selecting the decryption key.  The leftmost 64 bits of
the fingerprint are used (v4 uses the rightmost).

Aside from a few editorial changes the actual changes are:

* Revocation key and Issuer Fingerprint:

  - For a V5 key the 25 leftmost octets are used.

* Public key packet:

  - New four-octet count of the public key material.
    This is to ease parsing.

* Secret key packet

  - S2K Usage octet MUST NOT be 255.
    That is V5 keys require the SHA-1 checksum but
    we may want to drop this in favor of an AEAD mode.

  - New one-octet count of the S2K parameters.
    This is to ease parsing.

  - New four-octet count of the secret key material.
    This is to ease parsing.

* Key IDs and Fingerprint

  - The V5 fingerprint uses SHA-256

  - The magic header is changed from 0x99 + two-octets length header
    to 0x9a = four-octet length header.  The four-octet public key
    material count is inserted.

* EC DH Algorithm

  - For the 20 octets representing a recipient in the KDF parameters
    the v5 fingerprint truncated to 20 octets is used.