Deprecate legacy hash algorithms
MD5 has been deprecated for a long time; clearly deprecating this
algorithm using "SHOULD NOT implement" is thus due.

SHA-1 is still required to verify existing signature and can't be
deprecated.  However it is not anymore a mandatory algorithm with the
exception of MDC packets which we need to support at least read-only
for the foreseeable future.

Upgrading SHA2-256 to a mandatory algorithm should be obvious.

Keeping SHA2-512 optional benefits implementations on low end
platforms.