Proposed patch to add OCB to AEAD section
Hi openpgp WGers,

This is the proposed patch to add OCB to 4880bis.

--

<wk>
There has been a heated debate over the inclusion of a patented
algorithm into OpenPGP.  As it stands now we do not have a rough
consensus on adding this to 4880bis.  Instead it was suggested to put
this into a separate document.

For ease of working on this *I-D* I decided to include the patch
*for now*.  It will likely be removed before we finish the document.
</>