Clarify where a Revocation Key subpacket is to be placed.
Existing code always puts it into a 0x1F self-signature and thus
before the first user-id packet.  This makes sense because the scope
for a Revocation key is always the entire key.  Having it on a subkey
or a user-id would be surprising because it would still act on the
entire key.