Improved readability of the computing Signatures Section
This also moved the data from the Lietral Data packet into the final
trailer.  The reason for this is that implementaions might have
started to hash the data before they create the literal data packet.
For example if an unfinalized hash state is passed as input to the
signing implementation.

Note that a V5 signatures makes it impossible to convert a detached
signatures into a standard signature.