Add tools to the repo
It seems that the current tools availabale in Debian do not work as
expected.  I have always used a local installation of pandoc2rfc which
is now included in the repo for more consistent results.

I tested this on a new Devuan ASCII (based on Debian 9) and the only
change are lightly different page breaks.