Revert to the RFC4880 requirement of having a User ID.
With the below referenced patch Derek Atkins integrated his
"Device-Certificate Draft" to allow the use of a stripped down OpenPGP
key by space constrained devices.  The draft was never meant as a
general lifting of requirements which were intentionally introduced
when formalizing the old PGP 2 formats as OpenPGP.  This patch
clarifies this.

Fixes-commit: 3c3120a02d7b44621f3a7361ef75bdb7b7ade259
Signed-off-by: Werner Koch <wk@gnupg.org>