Introduce the Key Block subpacket to align OpenPGP with CMS.
This new subpacket may be used similar to the CertificateSet of
CMS (RFC-5652) and thus allows to start encrypted communication after
having received a signed message.  In practice a stripped down version
of the key should be including having only the key material and the
self-signatures which are really useful and shall be used by the
recipient to reply encrypted.