Fix signature packet description for EdDSA
The first change is that: In EdDSA, R is an EC point (not an integer
like ECDSA).  The (compressed) format of an EC point in EdDSA is
described at another part in the document.

The second change is that: In EdDSA, S is an integer.  An expression
"compressed" is not relevant here.  It is little endian in the native
semantics, but is put in MPI.  Well, the description of mine here is
still too short and ambiguous, which is needed fix later.