markdown-ize the original RFC 4880
This is done with an automated transformation of rfc4880.txt
(stripping pagebreaks and left-hand whitespace, and then some
additional manual cleanup.

Using the standard Make workflow in this repository, it produces an
almost-identical copy of rfc4880.txt, with a few exceptions:

 - bullets look like * instead of -

 - tables are laid out differently

 - no "Status of This Memo" or IETF Trust 2007 copyright statement

 - "Working group can be contacted via the current chair" is missing
 - Lutz Donnerhacke street address is missing

 - minor reflow changes based on padding, hyphenation, pagination.

 - page footers in .txt don't say "RFC 4880"

 - document header doesn't list obsoletes, RFC number, or category

None of these changes are relevant for the RFC 4880bis work, so this
should be workable as a source document.