Simplify and normalize AEAD text, in particular between EAX and OCB.
Outside of OCB- and EAX-specific sections, we already have a
requirement that the IV not be reused in AEAD.

IV reuse alone (without key reuse) isn't sufficient for a catastrophic
break in the crypto, because otherwise anyone could extract the IV
from an existing packet that they don't know the key for, and reuse it
with a different key to break the algorithm, so clarify what causes
the catastrophic failures.

Drop information about non-explicit IVs in OCB -- this is the AEAD
encrypted data packet section, so all IVs must be explicit.  When AEAD
shows up elsewhere, that's the location to specify how the IV is
chosen.

Add columns to the AEAD algorithm registry that indicate IV length and
tag length in octets.