Omit "starting" descriptor for "initialization vector"
The term "starting" seems redundant with "initialization".  And it
does not appear that there is any "non-starting" IV either.

For both EAX and OCB, there is a nonce that is generated from the IV,
but there are no other IVs to be concerned about, so the additional
word appears to introduce confusion instead of clarity.