Note collision between S2K usage octet and symmetric ciphers
These values have always collided, but we should be clearer about that
as we continue considering adding values to both ranges.

This makes the overlap (and potential conflict) explicit.

One way to resolve this is to limit the range of symmetric ciphers
that can be treated this way, instead of just "known cipher
algorithms".  Do we have a sense of which ciphers have ever been used
this way?