Summary table for secret key protection
The descriptions of secret key protection have always been difficult
to follow.  Hopefully, by presenting them in a summary form, it will
be easier for newer implementers to make sense of them.

Seeing them in tabular form makes me wonder whether we want to make an
IANA registry out of them, but i'm not sure.  It also makes me a
little nervous that we're sharing this codepoint space with symmetric
ciphers.