Simplify description of wire formats for elliptic curve points
The existing description of wire formats for elliptic curve points
was verbose, cumbersome, and overly flexible.

Introduce a simple table to summarize the wire formats, and add
subsection headings.

Note that this change introduces the idea that specific curves will
always use specific wire format representations, which was *not* the
case in earlier versions of the spec.

The OpenPGP design team agreed on this simplification, and no one
knows of any implementations that used an unusual point format for any
given curve.  At any rate, no signalling mechanism has ever been
available to indicate which EC point wire formats are understood for
which curves, so generators are likely to stick to existing patterns,
which this draft will document.