Add table representing value encodings, specifically to mention bytestrings
I wrote this table as specific to elliptic curves, because these hacks
to translate to and from an MPI are only used in ECC, and I wanted to
draw the analogy to the EC point representation formats.

I wrote that bytes(N) is "typically" fixed-length and didn't provide
any guidance of what to do for arbitrary-length strings (e.g. what to
do if the number of bits exceeds N*8) because i don't see any active
use of that in the specification as it stands.

Obviously, it would be possible to define an arbitrary-length string
in this way -- as long as it doesn't have any leading 0 octets -- by
saying that MPIs whose bitcounts exceed N*8 should just be accepted as
a raw octet string.  But i don't see any concrete use for that -- and
if we need an arbitrary-length octet string in the future, i'm wary
about using this structure (i'd rather use something else), because
not being able to represent any leading 0 octets could lead to
obscure problems.

This is close to what was called "SBS" in a previous attempt at
implementing this change.