Add prefixed octets representation
Like the octet string mechanism, this proposal hews to a fixed-length
set of octets because it seems unwise to include arbitrary-length
octet-strings if there's no way to represent a leading zero octet.

We arbitrarily select the specific prefix octet of 0x40, since there
are no other prefix octets contemplated for use in the specification.