Clarify when specific choices of S2K are appropriate.
This is an attempt to make it simpler to determine when to use
the different S2K approaches that OpenPGP offers.

In particular, it applies SHOULD NOT to the generation of Simple or
Salted S2K schemes unless the string itself is known to be
key-equivalent.

I was tempted to try to make Salted S2K a MUST NOT, in the hopes that
some future implementers can ignore at least one of the mechanisms,
but the existing section for SKESK (RFC 4880 §5.3) seems to imply that
Simple S2K is already a MUST NOT for SKESK.

I've opened
https://gitlab.com/sequoia-pgp/openpgp-interoperability-test-suite/-/issues/61
to try to figure out how much that SKESK Simple S2K "MUST NOT" is
really a meaningful interop issue.

I expect some of the text amended here to be updated if we end up
including Argon2, of course.