Document weird interaction between v5 SKESK and SEIPD
I think this captures the details pointed out by Daniel Huigens and
Justus Winter in
https://gitlab.com/openpgp-wg/rfc4880bis/-/merge_requests/53

Another approach would be instead to say that in the event that a v5
SKESK packet is supplied to unlock an SEIPD packet, the SEIPD packet
MUST use the same symmetric cipher used by in the v5 SKESK packet
itself.

But probably better to just explicitly forbid generation of that
combination so that maintainers don't have to think about it or
implement it.