Drop questionable advice about dropping keys.
Suppose we ever remove a subkey, then we have no place to put the
revocation certificates into. Even if we had, we could no longer
verify them, because we don't have the subkey anymore. If we now ever
consume a certificate update from a keyserver, and the server's copy
of the certificate lacks the revocation certificate, we may re-import
the subkey.

Therefore, I don't think that it is safe to ever remove a subkey, and
the RFC shouldn't suggest that.

dkg notes that the above reasoning is only true if the software
maintains the secret key in transferable form, which I think it
should (or at least be able to form the TSK for interoperability).  He
writes:

> Arguably, the thing that is most important to avoid losing is the
> key revocation signature (sigtype 0x20, aka "revcert") that
> invalidates it.  So it's conceivable that an implementation could
> decide to not store the subkey itself, as long as it keeps the revcert
> around, and is capable of re-applying the revcert to the subkey should
> it re-encounter it during a refresh of the certificate itself.
>
> However, doing that would require keeping the OpenPGP certificate in
> a form that is not just a Transferable Public Key, and it would only
> save a small amount of storage, since the subkey packet itself isn't
> likely to be much larger than 1KiB.
>
> So on balance, i don't think it's a good idea to suggest that an
> implementation should consider dropping a subkey, because it ends up
> dragging in a lot more complexity for minimal gain.

Fixes #39.