Generalize the mention of SED packets to all encryption containers.
Currently, the sections introducing PKESK and SKESK packets very
prominently mention the Symmetrically Encrypted Data packet, which is
odd because this is the one we don't want people to use.  Instead, use
the generic term 'encryption container' to refer to either the
Symmetrically Encrypted and Integrity Protected Data Packet, an AEAD
Encrypted Data Packet, or -for historic data- a Symmetrically
Encrypted Data Packet.