Indicate support for AEAD data (packet 20) in feature packet
This had originally been reserved for "AEAD Data and v5 SKESK" but
it's not clear to me why these need to be coupled together.

For one thing, an SKESK isn't formed with a particular OpenPGP
certificate in mind, so there is no feature flag to look at when
generating an SKESK -- it's not useful when deciding what kind of
SKESK to create.

Decisions about SKESK also revolve around choices of S2K algorithm,
which are not indicated in any features or preferences subpacket, so
simply indicating details about the version of SKESK support isn't
sufficient to guide parameter selection during SKESK generation.

Finally, if an implementation can handle an AEAD Data packet but not
v5 SKESK (for whatever reason), what should it indicate in this
feature flag?  It's unclear.

It seems cleaner and simpler to just make this by analogy with the MDC
-- this flag indicates to someone sending an encrypted message which
flavor of encrypted container is supported by the recipient.