Drop Features subpacket bit 0x04 -- it is not actionable
Without a clearer justification for why this bit should be present, it
doesn't make sense to include it in the specification.  It might even
cause interop failures (e.g. if someone with a v5 key decides not to
make a signature because their implementation doesn't see it
advertised on the recipient's certificate, even if the receipient can
handle the signature).

Moving it back to "reserved" lets us avoid creating problems and
confusing future implementers.

Closes: #50