Deprecate Revocation Key (prefer escrowed revocation signatures)
The Revocation Key subpacket is a tricky thing, and it's not clear
that it accomplishes anything that a simpler mechanism wouldn't also
accomplish.

(it'd also be nice to get more information about how well it's supported:
https://gitlab.com/sequoia-pgp/openpgp-interoperability-test-suite/-/issues/67)

Addresses part of #45, as this is one place where the fpr algo is
embedded in the standard.