Features subpacket AEAD flag should specify version 1
The AEAD packet has a version number.  Only version 1 is defined.

If version 2 is ever defined, it's not clear how an implementation
will signal that it supports the new version.

By making the feature flag describe the specific AEAD version
supported, we leave room to use an additional bit for each new
version.

We could probably retcon this whenever AEAD v2 gets defined, but it
seems safer to do it explicitly.