Clarify handling packet boundaries in a stream.
Be clear that the length indicators in the packet boundaries are the
canonical source for framing information.

The goal here is to be explicit about what to include in the packet,
and how to find the next packet in the stream, without being
prescriptive about what to do when internal size expectations don't
line up with what is explicitly described in the packet header.

See https://tests.sequoia-pgp.org/#Packet_excess_consumption