Key separation between different AEAD packet versions
Add the AEAD Encrypted Data packet tag and version to decrypted
session keys, and require the implementation to check them when
using a PKESK or SKESK to decrypt an AEAD Encrypted Data packet.

Also, clarify the allowed packet version combinations.

Also, clarify the input to the various public-key encryption
algorithms, and the padding used for each algorithm.