Clarify certification capabilities and primary keys.
Basically, parimary keys can always make self-signatures over their
own user IDs and subkeys, regardless of whether the "certification"
key flag is set.

It's conceivable at some point that we want to allow encryption-only
primary keys (i.e. a minimalist OpenPGP certificate for receiving
encrypted messages, where the user ID is somehow bound to the primary
key *without* a self-sig), but that would be a separate discussion.

This is an attempt to salvage usable ideas from !50.