Drop questionable advice on compression.
Drop the compression from the encryption workflow.  Not only is it
odd, but it is in the wrong place, after the encryption.

Drop the SHOULD compress the message.

Drop questionable advice about compression providing ciphertext
integrity.  EFAIL has shown us that the opposite is the case:
Compression can be used to hide artifacts of the attack.