Widen v5 Key IDs to 96 bits.
With the v5 fingerprints being generated using SHA256, they became
even more unwieldy than v4 fingerprints.

Whether we like it or not, comparing fingerprints is often the only
way to authenticate OpenPGP certs.  People print them on business
cards and hand them out, people compare them at key signing parties,
people include them in email footers.

Hex(SHA256(..)) is 64 characters, even without pretty printing.  That
is not only terrible for comparisons, it also creates problems for
simply displaying the values.

There are two properties to consider, both closely related to hash
algorithm properties:

- Collision resistance: What is the cost of producing two keys with
  the same Key ID?

  The setting is that a malicious entity creates two keys with
  colliding Key IDs.  While this may create some confusion, the
  confusion is only about keys that are both under the control of the
  malicious entity.  I cannot seem to imagine a way to turn that into
  an attack, and dkg seems to agree [0].

0: https://moderncrypto.org/mail-archive/messaging/2014/000013.html

- (Second) pre-image resistance: Given a Key ID (or a key),
  what is the cost of creating a second distinct key with the same
  Key ID?

  This is clearly a concern: If a malicious entity is able to
  fabricate a key matching the Key ID of a third party, then
  this malicious key may be mistaken for the real key.  This breaks
  authentication via Key ID comparison.

The design team has chosen SHA256 as the hash algorithm to compute the
fingerprints.  To improve ergonomics, we can render it differently
when asking the user to authenticate keys.  The obvious way is to
truncate the fingerprint to \(l\) bits.  Now the question is how that
truncation affects second pre-image resistance.  Fortunately, [1]
gives guidance: Truncating the digest of an approved hash
function (such as SHA2-256) to \(l\) bits results in an expected
second pre-image resistance of \(l\), i.e. it requires an exhaustive
search in the order of \(2^l\) computations of the hash function.

1: https://www.nist.gov/publications/recommendation-applications-using-approved-hash-algorithms

Thanks to its use in the Bitcoin network, we know tight limits on the
power required to compute SHA256 hashes [2].  Using this bound, we can
estimate that computing a (second) pre-image for a 96 bit Key ID
requires on average 31_100 years worth of the worlds electricity
production [3].

2: https://ieeexplore.ieee.org/document/9367201
3: https://ourworldindata.org/energy

Thus, a 96 bit Key ID can be used everywhere instead of a fingerprint
unless you require collision resistance.

Fixes #44.