Remove redundant octet count in v5 secret key packets.
The v5 secret key packet adds two octet counts, one for the S2K
parameters, and one for the secret key material.

Having an octet count for the S2K parameters helps to parse the packet
if the S2K method is unknown.

However, the secret key material octet count is redundant: the secret
key material extends to the end of the packet.  This information does
not add anything, but needlessly creates an error condition where the
octet count and the packet's remaining octets disagree.

Fixes #33.