v4 keys MUST make v4 sigs
This reduces the complexity of the spec:

When generating a signature, the implementation does not need to make
any decision about whether to make v4 signature or a v5 signature; it
falls out automatically from which key is used for signing.

This also means that there is no room for an implementation to ask the
user to make a decision that they are ill-equipped to make.

For a user that really wants to make a v5 signature, they can do so by
adopting a v5 key.