Rule out Argon2 with S2K usage octet 255
This is a safe requirement, as no compatible implementation would have
used Argon2 (specified in this draft) and alongside S2K usage octet
255 (which this draft says is MUST NOT generate).

A guaranteed rejection avoids any possible future trouble with key
reuse in CFB by fiddling with the two different S2K usage octets.

See
https://gitlab.com/openpgp-wg/rfc4880bis/-/merge_requests/128#note_823832961
for the origins of this discussion.