Change nonces to be monotonically increasing counters.
Previously, we computed the nonce by xoring an initialization vector
with the chunk counter.  This results in nonces that are not
monotonically increasing counters, which may prevent optimizations in
AEAD constructions.  A notable example of this is OCB when the API
assumes nonces to be a counter.

Instead, we build the nonces by taking fewer octets as IV, and
concatenating the chunk counter.  Now, the nonces are monotonically
increasing.

At first glance, this seems to reduce the entropy in the nonce, and
hence reduce the security of the scheme.  However:

- Nonces don't have to have a high entropy.  In fact, simple counter
  nonces are fine.  Note that we are improving on that by adding some
  IV bytes at the front.  This addresses the batch-attack concerns
  outlined here: https://blog.cr.yp.to/20151120-batchattacks.html

- The previous scheme was essentially a counter, but the counter-like
  nature was obscured by xoring in a static value.  It is not clear
  whether this is really an improvement, or merely obscures the
  monotonicity.

If this change turns out to be problematic, we can atomically revert
it before the publication of the final document.

Fixes #84.