Do not include the chunk index in the authenticated additional data.
Previously, we included the chunk index in the authenticated
additional data when encrypting chunks using AEAD.  However, this
prevents an optimization in AEAD modes where processing of the
additional data does not depend on the nonce, and thus the additional
data has to be processed just once independently of the number of
chunks.  Notably, OCB allows this.

Instead, we now no longer include the chunk index, therefore, the
additional data is constant, and can be processed just once.

At first glance, this seems to reduce the security of the
construction.  Note, however, that the chunk index is part of the
nonce, so any kind of reordering, dropping, or duplication of chunks
will result in the chunk being decrypted with the wrong nonce, making
the decryption operation fail.

If this change turns out to be problematic, we can atomically revert
it before the publication of the final document.

Fixes #86.