Further justification for padding with random octets
An OpenPGP implementation probably should not both pad and compress
the same information.  But sometimes, an OpenPGP object might be
subject to compression by external systems, and we want the padding to
not disappear in those cases either.

For example, for certificate transfer (e.g., over WKD) the certificate
might be compressed by a web server that provides automatic gzip
compression.

Such compression might still leak some packet length information: the
low-entropy parts of the certificate, like the User ID and some
subpackets might still be compressed.  But compared to the wide range
of different sizes of specific OpenPGP certificates, this is still a
better outcome (fewer anonymity sets, larger anonymity sets) even if
it is not perfect.