Replace Preferred AEAD Algorithms with Ciphersuites
This allows for more flexible specification of what ciphers and AEAD
modes an implementation supports, rather than forcing implementations
to support a full cross-product.

We allocate a different subpacket codepoint because of concerns raised
by @twiss1 that some of the older flavor of subpackets are already out
in the wild.

Closes: #94