Drop sourcecode markers from rendered variants.
Code that scans RFCs for extractable sourcecode snippets should be
scanning the XML variant, where the sourcecode is now properly
annotated (both with MIME type and with filename).

txt, html, and pdf renderings don't need the ugly <CODE BEGINS> and
<CODE ENDS> pseudo-tag as long as the code start and end are clearly
defined and can be copy/pasted.

In all rendered variants, the section headers clearly delimit the
code, and in the html and pdf variants, the code is even included
inside a gray box.