Enable robust parsing of v5 SKESK packets and encrypted keys.
In a42ea9c05479cc8b0424c85b865d796d2250586c we changed how KEKs are
computed, hence breaking all existing SKESKv5 packets.

This is an opportunity to fix a wart in the specification of the
SKESKv5 packet: it contains a number of variable-length fields, the
S2K object, the AEAD IV, and the encrypted session key with the AEAD
tag.  (The same problem exists in SKESKv4.)

- The S2K object's length is not explicitly given.  To parse the
  field, you need to look at the first octet, the tag, and then use a
  look-up table.

  The problem are unknown S2K types, either private ones, or
  not-yet-defined ones.

- The length of the IV field depends on the AEAD mode.  Again, if the
  AEAD mode is not known, you don't know how many octets to parse.

- The encrypted session key's length can be determined because we know
  the length of the packet and all other field lengths.

- The length of the AEAD tag is also dependent on the AEAD mode, but
  knowing its length is not as important because it can be seen as
  part of the encrypted session key.

Now, you may argue that this is a cosmetic fix: why would you want to
know the encrypted session key if you don't understand the S2K
mechanism or AEAD modeused to encrypt it.  If you don't implement the
algorithms, what good is looking at the individual fields?

My answer to that is modeling: If you are able to parse a packet, even
if you don't understand or implement a part of it (like the S2K
mechanism), you can model that part as unknown.  Explicitly modeling
unknown values leads to gracefully degrading operations.  In our
experience that leads to robustness in the face of malformed data and
forward compatibility.

An example of this kind of robustness (even if in a different setting)
is an encrypted message with two recipients using different versions
of OpenPGP.  The older implementation must gracefully ignore the PKESK
packet it doesn't understand.

So from that perspective, it is not merely a cosmetic fix.

On the other hand, it doesn't change the semantics in any way, and I
don't think it introduces a new error condition: if the S2K object's
length differs from the expected length (previously implicit, now
explicit), parsing the packet will fail or produce unexpected results.

The fix enables robust parsing of v5 SKESK packets and v5 encrypted
key packets containing unknown S2Ks or AEAD modes.  It also harmonizes
the wire format of v5 SKESK packets and v5 encrypted key packets.