Avoid incorrect secret key S2K parameter field count.
Avoid specifying the number of S2K parameter fields, since we've
gotten it wrong repeatedly.

Split out the summary table into v4 and v5 summaries, to accomodate
the new octet counts clearly.

Note that this might have a minor functional change compared to the
prior text for cleartext (unencrypted) secret keys.  The prior text
*always* included a one-octet length count for the S2K params, but
this variant omits the length count when the key is in the clear.

Closes: #32