Drop redundant and problematic paragraph.
Why does the primary key collect certifications?  Usually,
certifications are over (primary key, user {id, attribute}) pairs.
It *makes* certifications.

I don't think we should use the term "certifying" for binding of
subkeys.  I think consensus is that a primary key that is not marked
as certification-capable can still bind user ids and subkeys.

All in all, what the paragraph is trying to say is better said in
the one before.