Tweak TSK description to allow for missing secret key material.
Sometimes, a TSK has to be transferred even if not all secrets are
known or should be transferred.  Notable examples are "offline primary
keys", subkeys with secrets on a smart card, or additional archival
encryption keys.

The current practice is to use a secret-key packet, but include a stub
secret key.  This seems unfortunate, because a) it has not been
standardized, and b) this seems to be a modeling error: it is a
secret-key packet with only public-key material, and should thus be a
public-key packet.

Fixes #52.