Document "strict" packet type grammars
A strict grammar has the benefit that it is harder to craft a packet
sequence that might (for example) decrypt to one cleartext in one
implementation, and a different cleartext in a different
implementation.

Encouraging implementations to be strict in their grammars should
avoid this kind of interoperability failure.

In the next step, we will partition the unallocated packet space into
critical and non-critical packets.  This way, making the grammar
strict doesn't lose any potential future flexibility.

Closes: #121