Re-frame "OpenPGP CFB mode" in terms of standard CFB
In a Symmetrically Encrypted Data packet, the CFB context is
"resynchronized" after encrypting the random prefix. This is
equivalent to creating a new CFB context with the last block-
size octets of the encrypted random prefix as the IV. Thus,
this patch re-frames it in those terms, so that we don't have
to specify a special "OpenPGP CFB mode", and that we can refer
to standard CFB instead.